+++ zaino-state/src/chain_index/tests.rs
@@ -1,3 +1,4 @@
 //! Zaino-State ChainIndex unit tests.
 
 mod finalised_state;
+mod vectors;
