+++ zaino-state/src/chain_index/tests/finalised_state/migrations.rs
@@ -1,0 +1,1 @@
+//! Holds database migration tests.
